+++ tests/cases.js
@@ -34,6 +34,7 @@ module.exports = [
 
 	'--test-entry super/i-block/test',
 	'--test-entry super/i-block/modules/dom/test',
+	'--test-entry super/i-block/modules/opt/test',
 	'--test-entry super/i-block/modules/block/test',
 	'--test-entry super/i-block/modules/field/test',
 	'--test-entry super/i-block/modules/storage/test',